型の定義を一つにまとめた
