--- src/types/date.ts
@@ -1,2 +1,0 @@
-// 日付を表す型
-export type DateType = Date | string;
